requirements_small.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
colorama==0.4.3
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
imageio==2.6.1
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.1
open3d-python==0.7.0.0
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
plydata==0.3.3
plyfile==0.7.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
pycodestyle==2.5.0
Pygments==2.5.2
pykdtree==1.3.1
pylint==2.4.4
pyparsing==2.4.6
pypng==0.0.20
pyrealsense2==2.32.1.1299
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
trimesh==3.5.24
typed-ast==1.4.1
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==2.2.0
